Reorganize project into lib/ folder structure
- Move all modules into organized lib/ subfolders:
  - lib/lap/ - Lap data structure, CSV parsing/export
  - lib/ui/ - Theme, utils, markdown, wedge
  - lib/controls/ - UI controls (corner, delta_bar, lap_picker, etc.)
  - lib/core/ - Core utilities (scoring, history, files, brake)
  - lib/state.lua - Centralized state
  - lib/settings.lua - App settings

- Remove LÖVE2D fallback from sim/ - ImGui only
- Update all require paths throughout codebase
- Update test files for new paths

Structure:
  ac-tracer.lua (entry point)
  lib/
    lap/init.lua, csv_parser.lua, csv_export.lua
    ui/theme.lua, utils.lua, markdown.lua, wedge.lua
    controls/corner.lua, corner_analysis.lua, delta_bar.lua, ...
    core/scoring.lua, history.lua, files.lua, brake.lua
    state.lua, settings.lua
  sim/main.lua, csp_shim.lua
