+++ tests/test_lap_picker.lua
@@ -3,24 +3,24 @@
 suite("lap_picker")
 
 test("refresh calls file_utils.invalidateCache", function()
-    local originalFileUtils = package.loaded['file_utils']
-    local originalLapPicker = package.loaded['lap_picker']
+    local originalFileUtils = package.loaded['lib.core.files']
+    local originalLapPicker = package.loaded['lib.controls.lap_picker']
 
     local calls = 0
-    package.loaded['file_utils'] = {
+    package.loaded['lib.core.files'] = {
         invalidateCache = function() calls = calls + 1 end,
         formatFileSize = function() return "1KB" end,
         formatLapTime = function() return "0:00.00" end,
         scanCSVFiles = function() return {} end,
     }
 
-    package.loaded['lap_picker'] = nil
-    local lap_picker = require('lap_picker')
+    package.loaded['lib.controls.lap_picker'] = nil
+    local lap_picker = require('lib.controls.lap_picker')
 
     lap_picker.refresh()
     assert_equal(calls, 1)
     assert_true(lap_picker.isLoading() == false)
 
-    package.loaded['file_utils'] = originalFileUtils
-    package.loaded['lap_picker'] = originalLapPicker
+    package.loaded['lib.core.files'] = originalFileUtils
+    package.loaded['lib.controls.lap_picker'] = originalLapPicker
 end)
